requirements.txt
antlr4-python3-runtime==4.7.2
black
cfpq-data==1.0.2
pre-commit
pydot
pyformlang
pytest
scipy
